WAL: Fix nasty file closing bug
When we use os.Create for the purpose of creating a file in the filesystem, we don't necessarily want to do anything with the returned file pointer. Therefore, we might think we can just not capture the returned file pointer by using '_'. However, we still need to *close* the file, as os.Create actually opens the file and keeps it open until we close it, whether or not we capture file pointer.

This bug was caught in the DeleteOldSegments function when renaming WAL logs. Since a previous append may have created a new segment, the associated opening of the file would never be closed, failing the os.Rename function as it cannot rename an opened file.