WAL: Change how DeleteOldSegments works
Previously we would use the lowest water mark as the index before which we would delete logs, however this showed to be impractical, as the amount of logs being deleted was too low and misleading.
Lowest water mark index now represents the *amount* of logs that we *keep* after running DeleteOldSegments.