WAL: Fix how DeleteOldSegments works
Previously, using a LWM of zero would crash the program. We now add a check to see if the segmentPaths slice has been completely emptied by the segment deletions and act accordingly.
NOTE: Because of how the WAL is implemented, after the initial New, it is not possible to have an empty log directory. Therefore, both DeleteOldSegments (when LWM is 0) and DeleteAllSegments have the added behavior of creating a new, empty segment that the WAL can use for subsequent appends. Let me know if this behavior is undesirable.

Closes #40