Memtable: Fix memusage summing
Since records can be of different sizes, and since adding records overrides potentially existing records with the same key, we need to account for this potential size difference when summing memusage.
This is the simplest fix, however the cleanest would be if skiplist's Write returned the previously existing record as well as its presence boolean. Then we wouldn't need to call Find.

Closes #54